fix: move function add_new_trainer from booking_crud.py to trainer_crud.py
